Stopped Zombies and Skeleton Warriors from walking off ledges
Modified the OnMoveTowardsPlayer behavior to check for ledges using Raycast2D's. If no collider is detected, it sets the velocity to 0 and sets a new _frozen bool to true, which causes all calls to MoveLeft() or MoveRight() to be passed over. However it also allows for the checking of player position to continue so when the enemy is able to move towards the player again, it will resume doing so. Currently the raycast offset must be modified for each new enemy so that the raycast doesnt collide with any of the enemy's own colliders, however I will work on filtering out these colliders. Theoretically, any enemy which utilizes this animation behavior should now no longer walk off of ledges. I will work on enemies which dont use this behavior, such as the dashing slime, next