PlayerPrefs sound control
Implemented changing music volume through sliders in the pause menu manager. To avoid conflicts with modifying the scene I did not include the two sliders in this commit. To implement them, add a On Value Changed action to each of them w/ the public function SaveMusicVolume or SaveSoundVolume. Currently changing the sound slider does nothing and the music slider really only affects the starting music. This is because a lot of the volume control seems hard coded with public values, so it will take some time to go through and modify each of those, but currently I do not have time to do that but will soon. Also, currently setting the volumes in PlayerPrefs is done w/ the PauseMenuManager object and script, but this can be changed easily once there is a settings menu of a sort.